refactor(wash-runtime): address P3 cross-component review feedback
- Extract linked-call machinery into engine/linked_call.rs and volume
  resolution into engine/volumes.rs; move param/result lowering into value.rs
- Resolve component volume mounts eagerly at resolve() (fail at deploy, not
  first request)
- Standardize workload.rs on wasmtime::error::Context;
- Document ComponentCtxTemplate/EphemeralLinkedCall/expand_link_closure; add
  ephemeral call traces
- Tidy fixture/test docs
