+++ crates/wash-runtime/tests/fixtures/res-caller-p3/src/lib.rs
@@ -30,7 +30,7 @@ impl Handler for Component {
         let headers = Fields::new();
         let (mut tx, rx) = bindings::wit_stream::new::<u8>();
         let (trailers_tx, trailers_rx) = bindings::wit_future::new(|| todo!());
-        wit_bindgen::spawn(async move {
+        wit_bindgen::spawn_local(async move {
             tx.write_all(body.into_bytes()).await;
             drop(tx);
             let _ = trailers_tx.write(Ok(None)).await;
